--- lerna-debug.log
@@ -1,10 +1,0 @@
-0 silly argv { _: [ 'run' ],
-0 silly argv   lernaVersion: '3.20.2',
-0 silly argv   '$0': '/Users/cmolloy/Documents/repos/integration/node_modules/.bin/lerna',
-0 silly argv   script: 'build' }
-1 notice cli v3.20.2
-2 verbose rootPath /Users/cmolloy/Documents/repos/integration
-3 error JSONError: Unexpected token < in JSON at position 40 while parsing near '.../integration-core",<<<<<<< HEAD  "vers...' in packages/integration-core/package.json
-3 error     at module.exports (/Users/cmolloy/Documents/repos/integration/node_modules/parse-json/index.js:26:19)
-3 error     at parse (/Users/cmolloy/Documents/repos/integration/node_modules/load-json-file/index.js:15:9)
-3 error     at pify.then.data (/Users/cmolloy/Documents/repos/integration/node_modules/load-json-file/index.js:18:94)
