+++ package.json
@@ -1,6 +1,6 @@
 {
-  "description": "Redhat UXD Integration Components",
-  "repository": "https://github.com/rh-uxd/integration.git",
+  "description": "Redhat UXD Application Services PatternFly Components",
+  "repository": "https://github.com/rh-uxd/appservices-patternfly.git",
   "license": "MIT",
   "private": true,
   "workspaces": {
