+++ packages/appservices-patternfly-core/src/common/IntegrationProductInfo.ts
