+++ packages/appservices-patternfly-core/styles/assets/Logo-Red_Hat-Integration-A-Reverse-RGB.png
