+++ packages/appservices-patternfly-core/styles/assets/Logo-Red_Hat-Managed_Integration-A-Reverse-RGB.png
