+++ packages/appservices-patternfly-core/webpack.config.js
