+++ packages/appservices-patternfly-crossconsole/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@rh-uxd/integration-react",
+  "name": "@rh-uxd/appservices-patternfly-crossconsole",
   "version": "1.0.3",
   "description": "Set of common react integration components",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/rh-uxd/integration.git"
+    "url": "git+https://github.com/rh-uxd/appservices-patternfly.git"
   },
   "main": "dist/js/index.js",
   "module": "dist/esm/index.js",
@@ -22,7 +22,7 @@
   "dependencies": {
     "@patternfly/react-core": "3.131.4",
     "@patternfly/react-icons": "3.14.30",
-    "@rh-uxd/integration-core": "1.0.3",
+    "@rh-uxd/appservices-patternfly-core": "1.0.3",
     "react": "16.13.1",
     "react-dom": "16.13.1"
   },
@@ -50,7 +50,7 @@
     "access": "public"
   },
   "bugs": {
-    "url": "https://github.com/rh-uxd/integration/issues"
+    "url": "https://github.com/rh-uxd/appservices-patternfly/issues"
   },
-  "homepage": "https://github.com/rh-uxd/integration#readme"
+  "homepage": "https://github.com/rh-uxd/appservices-patternfly#readme"
 }
