+++ packages/appservices-patternfly-crossconsole/src/components/CrossNavContextSelector/CrossNavContextSelectorConstants.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { CrossNavApp } from '@rh-uxd/integration-core';
+import { CrossNavApp } from '@rh-uxd/appservices-patternfly-core';
  
 export const ContextSelectorContext = React.createContext({
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
