+++ packages/appservices-patternfly-crossconsole/src/components/CrossNavContextSelector/index.ts
@@ -1,3 +1,3 @@
 export * from './CrossNavContextSelector'
 export * from './CrossNavContextSelectorItem'
-export { CrossNavApp, CrossNavAppState } from '@rh-uxd/integration-core' 
+export { CrossNavApp, CrossNavAppState } from '@rh-uxd/appservices-patternfly-core'
