+++ packages/appservices-patternfly-crossconsole/webpack.config.js
