+++ packages/demo-app-2/src/App.tsx
@@ -29,14 +29,14 @@ import {
   ToolbarGroup,
   ToolbarItem
 } from '@patternfly/react-core';
-import {CrossNavHeader, CrossNavApp, LoadingPage} from '@rh-uxd/integration-react';
+import {CrossNavHeader, CrossNavApp, LoadingPage} from '@rh-uxd/appservices-patternfly-crossconsole';
 // make sure you've installed @patternfly/patternfly
 import accessibleStyles from '@patternfly/react-styles/css/utilities/Accessibility/accessibility';
 import spacingStyles from '@patternfly/react-styles/css/utilities/Spacing/spacing';
 import { css } from '@patternfly/react-styles';
 import { BellIcon, CogIcon } from '@patternfly/react-icons';
 import imgAvatar from './logo.svg';
-import rhIntegrationLogo from './Logo-Red_Hat-Integration-A-Reverse-RGB.png';
+import rhIntegrationLogo from './Logo-Red_Hat-Application Services PatternFly-A-Reverse-RGB.png';
 //import './App.css'
 
 export class App extends React.Component<{}, {
@@ -186,7 +186,7 @@ export class App extends React.Component<{}, {
         <CrossNavHeader
           apps={apps}
           currentApp={{id: 'second-demo-app', name: 'Second Demo App', rootUrl:'localhost:3001', isHttp: true}}
-          logo={<Brand src={rhIntegrationLogo} alt="Red Hat Integration Logo" />}
+          logo={<Brand src={rhIntegrationLogo} alt="Red Hat Application Services PatternFly Logo" />}
           toolbar={PageToolbar}
           avatar={<Avatar src={imgAvatar} alt="Avatar image" />}
         />
