+++ packages/demo-app/package.json
@@ -5,8 +5,8 @@
   "dependencies": {
     "@patternfly/react-core": "3.129.3",
     "@patternfly/react-icons": "3.14.30",
-    "@rh-uxd/integration-react": "1.0.3",
-    "@rh-uxd/integration-core": "1.0.3",
+    "@rh-uxd/appservices-patternfly-crossconsole": "1.0.5",
+    "@rh-uxd/appservices-patternfly-core": "1.0.5",
     "@testing-library/jest-dom": "5.9.0",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "10.4.1",
